refactor(daemon): split things to channel and memory
