--- app/daemon/src/channel/mod.rs
@@ -1,3 +1,0 @@
-//! Channel routing and management.
-
-pub mod router;
