+++ app/daemon/src/lib.rs
@@ -1,13 +1,12 @@
 //! Walrus gateway — application shell composing runtime, channels, and cron
 //! scheduling. Personal agent, local-first.
 
-pub mod channel;
 pub mod config;
 pub mod feature;
 pub mod gateway;
 pub mod loader;
 
-pub use channel::router::{ChannelRouter, RoutingRule};
+pub use channel::{ChannelRouter, RoutingRule};
 pub use config::DaemonConfig;
 pub use feature::cron::{CronJob, CronScheduler};
 pub use gateway::{
