+++ crates/memory/src/sqlite/memory.rs
@@ -1,7 +1,7 @@
 //! Memory trait implementation for SqliteMemory.
 
-use crate::SqliteMemory;
-use crate::sql;
+use crate::sqlite::SqliteMemory;
+use crate::sqlite::sql;
 use crate::utils::now_unix;
 use crate::{Embedder, Memory, MemoryEntry, RecallOptions};
 use anyhow::Result;
